Added seperate method for retrieving provinces, added empty default value
The empty default value reflects the same behavior as is included by default in GravityForms. When hiding a province field for the frontend user, the province would still be submitted by email and thus, included in the {all_fields} merge tag, because there was no option to select an empty default.